- m_bmask.c: fixed compile warning
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid/branches/8.2.x@10148 82007160-df01-0410-b94d-b575c5fd34c7

+++ modules/core/m_bmask.c
@@ -61,7 +61,7 @@ ms_bmask(struct Client *source_p, int parc, char *parv[])
   const char *mask;
   char *s, *t, *mbuf, *pbuf;
   dlink_list *list = NULL;
-  int mlen = 0, tlen = 0;
+  size_t mlen = 0, tlen = 0;
   int modecount = 0;
   unsigned int type = 0;
